+++ scripts/serve/serve-static-cli.js
@@ -10,6 +10,7 @@ commander
   .version(packageJson.version)
   .option('--config <path>', 'The webpack config to serve. Alias for <config>.', undefined)
   .option('--port <n>', 'The port the app should listen on', parseInt)
+  .option('--host <s>', 'Sets the host that the server will listen on. eg. \'10.10.10.1\'', undefined)
   .option('-p, --production', 'Passes the -p flag to the webpack config, if available.')
   .option('--site <path>', 'The relative path to the static site. This takes precidence over webpack config if both are passed.', undefined)
   .option('--disk', 'The webpack assets will be written to disk instead of a virtual file system.')
@@ -19,8 +20,9 @@ const port = commander.port || process.env.PORT;
 
 serve({
   config: loadDefaultWebpackConfig(commander.config),
-  site: commander.site,
   disk: commander.disk,
+  host: commander.host,
   port,
   production: commander.production,
+  site: commander.site,
 });