+++ scripts/serve/serve.js
@@ -1,15 +1,18 @@
 const serve = require('webpack-serve');
 
 const webpackServe = (options) => {
+  const { port, host } = options;
   let config = options.config;
   if (typeof config === 'function') {
     config = config(undefined, { p: options.production });
   }
 
-  const serveConfig = { config };
-  if (options.port) {
-    serveConfig.port = options.port;
-  }
+  const serveConfig = {
+    config,
+    ...(port) && { port },
+    ...(host) && { host },
+  };
+
   serve(serveConfig);
 };